Add dual_basis and dual_evaluation to elements
Co-authored-by: Matthew Kan <ting.kan19@imperial.ac.uk>

A new dual_basis property is added to all elements which represents the
dual basis as a (generally sparse) gem weight tensor Q and corresponding
point set x.
The general dual evaluation is then Q * fn(x) (the contraction of Q
with fn(x) along the the indices of x and any shape introduced by fn).

The the dual_evaluation method returns a gem expression for the dual
evaluation contraction given some function fn.

For more, see the relevant docstrings