Create QuadratureElement from ref cell and rule
This changes the FInAT QuadratureElement API to have it be created from
a FIAT reference cell and FInAT quadrature rule
(AbstractQuadratureRule). This allows some extra checks to be added when
making a QuadratureElement and generally makes the API a bit more
robust.

To create an element the old way introduce a new make_quadrature_rule
function which takes a FIAT reference cell, quadrature degree and
optional scheme.

Also corrects a small documentation error.