QuadratureElement FIAT equivalent with unknown points
Pass array of NaNs as points when creating FIAT equivalent to
QuadratureElement.

This really should raise a ValueError since there ought not to be
a FIAT equivalent where the points are unknown.

For compatibility until FInAT dual evaluation is done we pass an
array of NaNs of correct shape as the points.