Don't reset all warning filters on element construction
I guess the code here was trying to make any warnings related to the
enclosed LU factorisation/solve into errors, but it was then resetting
the entire warnings filter. This shouldn't be done from library code,
the pattern should be to use the catch_warnings context manager to
ensure the previous state is restored.